Add support for :raw flag to api and GET, etc
:raw true causes GET and friends to not destructure, coerce, or otherwise
fiddle with the request, except as specified in the compojure
request-destructuring clause - things like :query-params produce swagger
metadata, but do *not* cause any parameters to be destructured out of the
request.

Additionally, (api {:raw true ...} handler) support is added, which prevents
wrapping handler in the compojure-api middlewares.

Proposed implementation of the feature(s) requested in #240.